Moved Symbol and SymbolFactory to 'wyopcl.translator.bound' package.
